refactor(path_planners): share grid setup and RRT run initialization
Give A* and JPS a common GridPlannerBase for bounds, resolution, and grid
shape, and move the RRT family's per-run start/goal and tree reset into
RRT._init_search. Remove the unused RRT._plot_circle and the commented-out
code in PRM.
